refactor: scss, styled-component 리팩토링
page 전용 scss를 따로 만들었다.
