+++ client/src/components/Loading.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import styled from 'styled-components';
 
-const LoadingWrapper = styled.div`
+const LoadingContainer = styled.div`
   width: 100%;
   height: 100%;
-  min-height: 1120px;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -13,9 +12,9 @@ const LoadingWrapper = styled.div`
 
 const Loading = () => {
   return (
-    <LoadingWrapper>
+    <LoadingContainer>
       <CircularProgress />
-    </LoadingWrapper>
+    </LoadingContainer>
   );
 };
