+++ client/src/pages/mainPage.jsx
@@ -3,18 +3,17 @@ import Header from '../components/header/Header';
 import MainSection from '../components/main-section/MainSection';
 import SideSection from '../components/side-section/SideSection';
 
-import style from '../styles/makePage.module.scss';
-
+import style from '../styles/scss/Page.module.scss';
 
 const MakePage = () => {
-    return (
-        <div className={style.makePageContainer}>
-          <Header />
-          <div className={style.sectionWrapper}>
-            <MainSection />
-            <SideSection />
-          </div>
-        </div>
-    );
-}
+  return (
+    <div className={style.PageContainer}>
+      <Header />
+      <div className={style.SectionWrapper}>
+        <MainSection />
+        <SideSection />
+      </div>
+    </div>
+  );
+};
 export default MakePage;
