+++ client/src/styles/scss/Page.module.scss
@@ -1,0 +1,14 @@
+@import './global.scss';
+
+.PageContainer {
+  width: 100%;
+  height: 100%;
+}
+
+.SectionWrapper {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
