+++ client/src/styles/scss/global.scss
@@ -1,0 +1,43 @@
+@mixin FlexBox($direction: row, $justify-content: center, $align-items: center) {
+  display: flex;
+  flex-direction: $direction;
+  justify-content: $justify-content;
+  align-items: $align-items;
+}
+
+$white: #ffffff;
+$off-white: #f6f6f6;
+$grey1: #888888;
+$grey2: #bbbbbb;
+$grey3: #d7d7d7;
+$title-active: #222222;
+$blue: #0055fb;
+$red: #f45452;
+$primary: #ffc700;
+$primary-dark: #e8a20c;
+
+@font-face {
+  font-family: 'Noto Sans KR';
+  font-weight: 300;
+  src: url('../fonts/NotoSansKR-Light.otf') format('woff2'), url('../fonts/NotoSansKR-Light.otf') format('woff'),
+    url('../fonts/NotoSansKR-Light.otf') format('truetype');
+}
+
+@font-face {
+  font-family: 'Noto Sans KR';
+  font-weight: 400;
+  src: url('../fonts/NotoSansKR-Regular.otf') format('woff2'), url('../fonts/NotoSansKR-Regular.otf') format('woff'),
+    url('../fonts/NotoSansKR-Regular.otf') format('truetype');
+}
+@font-face {
+  font-family: 'Noto Sans KR';
+  font-weight: 500;
+  src: url('../fonts/NotoSansKR-Medium.otf') format('woff2');
+}
+
+@font-face {
+  font-family: 'Noto Sans KR';
+  font-weight: 700;
+  src: url('../fonts/NotoSansKR-Bold.otf') format('woff2'), url('../fonts/NotoSansKR-Bold.otf') format('woff'),
+    url('../fonts/NotoSansKR-Bold.otf') format('truetype');
+}
