+++ client/src/styles/scss/index.scss
@@ -1,0 +1,20 @@
+@import './global.scss';
+
+* {
+  box-sizing: border-box;
+  font-family: 'Noto Sans KR', sans-serif;
+  margin: 0;
+  padding: 0;
+}
+
+html,
+body {
+  width: 100%;
+  height: 100%;
+}
+
+#root {
+  background: $off-white;
+  display: flex;
+  justify-content: center;
+}
